Fixed issue where the bounding boxes in video annotations were not saved: added functionality to add/remove keyframes when propagating the annotations.
The video annotation model was missing to key parameters: height and width. These have been added.

Refactored the functionality in annotationweb.js to remove keyframes to a dedicated function, so it can be used in other applications.

Change the destination URL when saving video annotations, to the correct function within the video annotations application.